pyproject.toml
[project]
name = "nba-props-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "certifi>=2025.8.3",
    "lightgbm>=4.6.0",
    "nba-api>=1.10.0",
    "pandas>=2.3.2",
    "psutil>=7.0.0",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "selenium>=4.35.0",
    "tqdm>=4.67.1",
    "xgboost>=3.0.5",
]
